% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_Data.R
\docType{data}
\name{emoji_Data}
\alias{emoji_Data}
\title{Emoji Data}
\format{
Dataframe with columns:
\describe{
 \item{C1,C2,C3,C4,C5,C6,C7,C8}{Uni-code in text.}
 \item{Browser}{Code applicable on Web Browser.}
 \item{Appl}{Code applicable on Apple.}
 \item{Goog}{Code applicable on Google.}
 \item{FB}{Code applicable on FaceBook.}
 \item{Wind}{Code applicable on Windows Devices.}
 \item{Twtr}{Code applicable on Twittwer.}
 \item{Sams}{Code applicable on Samsung.}
 \item{Gmail}{Code applicable on Gmail.}
 \item{Joy,SB,DCM,KDDI}{Code applicable on other Platforms.}
 \item{Description}{Code description}
}
}
\source{
<https://unicode.org/emoji/charts/full-emoji-list.html>
}
\usage{
emoji_Data
}
\description{
Emoji Data
}
\keyword{datasets}
