% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOI_News_Dataset.R
\name{TOI_News_Dataset}
\alias{TOI_News_Dataset}
\alias{TOI_News_Links}
\title{Creates Interim Dataset}
\usage{
TOI_News_Dataset(keywords)

TOI_News_Links(keywords)
}
\arguments{
\item{keywords}{A String, user-defined.}
}
\value{
Returns DataFrame based on \code{keywords}
}
\description{
Creates an interim news dataset based on user-defined keywords
 for all possible links extracted from respective source.
}
\section{Functions}{
\itemize{
\item \code{TOI_News_Links}: Extracts Source Links
}}

\examples{
#### Creates Dataset based on keysword
\donttest{
NewsData = TOI_News_Dataset("AirLines")
}
}
\seealso{
\code{\link{TOI_News_Articles}}
}
\author{
Vatsal Aima,  \email{vaima75@hotmail.com}
}
