% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOI_News_Articles.R
\name{TOI_News_Articles}
\alias{TOI_News_Articles}
\title{Extract Media News}
\usage{
TOI_News_Articles(keywords, IsDataFrame = F, IsDate = F, start_date, end_date)
}
\arguments{
\item{keywords}{A String, user-defined.}

\item{IsDataFrame}{Boolean Value, to determine whether the outcome should be
a Dataframe or files written to disk. if set to \emph{TRUE} then retuns
DataFrame (default \emph{FALSE}).}

\item{IsDate}{Boolean Value, to determine whether to perform filtering
operations based on dates. If set to \emph{TRUE} then Date operations will
be applied (default \emph{FALSE}).}

\item{start_date}{Date Value, provide the starting date \strong{FROM} where
the data should be extracted. NOTE: only provide \code{start_date} when
\code{IsDate} is set \emph{TRUE}}

\item{end_date}{Date Value, provide the ending date \strong{TO} where the
data should be extracted. \strong{\emph{NOTE}}: only provide
\code{end_date} when \code{IsDate} is set \emph{TRUE}}
}
\value{
Returns DataFrame or write files to the disk based on \code{keywords}
}
\description{
Creates a DataFrame or Write files disk by extracting text data
  from source and interim dataset based on user's keywords.
}
\examples{
#### Creates Dataset by filtering 31 days from current date
\dontrun{
NewsDataset1 = TOI_News_Articles(keywords = "Politics In US",
IsDataFrame = TRUE,
IsDate = TRUE,
start_date = Sys.Date()- 31,
end_date = Sys.Date())

# Creates Dataset by custom filtering through dates
NewsDataset2 = TOI_News_Articles(keywords = "BaseBall",
IsDataFrame = TRUE,
IsDate = TRUE,
start_date = "2019-09-20",
end_date = "2019-10-20")

# Write files to disk
TOI_News_Articles(keywords = "Goibibo")
}
}
\seealso{
\code{\link{TOI_News_Dataset}}
}
\author{
Vatsal Aima,  \email{vaima75@hotmail.com}
}
